Uploading project to GitHub
I'm uploading my Taiwanese Dictionary project to GitHub. I haven't touched this since 2014. Half the code is deprecated and I had to fiddle with some settings to get it to compile again.

I used ParseKit (which is/was a Pod) probably for regular expressions - something about the Pod setup means we have to open the project by opening 'Taiwanese.xcworkspace', otherwise something breaks.